Gather public types into one file
