+++ src/components/publicApi/UpgradeUsernameScreen.tsx
@@ -1,9 +1,9 @@
 import { EdgeAccount, EdgeContext } from 'edge-core-js'
 import * as React from 'react'
 
-import { asExperimentConfig, OnLogEvent } from '../../types/ReduxTypes'
 import { Router } from '../navigation/Router'
 import { ReduxStore } from '../services/ReduxStore'
+import { asExperimentConfig, OnLogEvent } from './publicTypes'
 
 interface Props {
   account: EdgeAccount
